# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ballistic Calculator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Drag model used for BC
 */
export type ProfileBcModel =
  (typeof ProfileBcModel)[keyof typeof ProfileBcModel];

export const ProfileBcModel = {
  G1: "G1",
  G7: "G7",
} as const;

export interface Profile {
  id: number;
  name: string;
  caliber: string;
  /** Bullet weight in grains */
  bulletWeight: number;
  /** Bullet diameter in inches */
  bulletDiameter: number;
  /** Muzzle velocity in feet per second */
  muzzleVelocity: number;
  /** Ballistic coefficient (G1 or G7) */
  ballisticCoefficient: number;
  /** Drag model used for BC */
  bcModel: ProfileBcModel;
  /** Zero range in yards */
  zeroRange: number;
  /** Scope height above bore in inches */
  scopeHeight: number;
  /** Rifle weight in pounds (for recoil calc) */
  rifleWeight: number;
  notes?: string | null;
  createdAt: string;
}

export type CreateProfileInputBcModel =
  (typeof CreateProfileInputBcModel)[keyof typeof CreateProfileInputBcModel];

export const CreateProfileInputBcModel = {
  G1: "G1",
  G7: "G7",
} as const;

export interface CreateProfileInput {
  name: string;
  caliber: string;
  bulletWeight: number;
  bulletDiameter: number;
  muzzleVelocity: number;
  ballisticCoefficient: number;
  bcModel: CreateProfileInputBcModel;
  zeroRange: number;
  scopeHeight: number;
  rifleWeight: number;
  notes?: string | null;
}

export type CalculationInputBcModel =
  (typeof CalculationInputBcModel)[keyof typeof CalculationInputBcModel];

export const CalculationInputBcModel = {
  G1: "G1",
  G7: "G7",
} as const;

export type CalculationInputUnitSystem =
  (typeof CalculationInputUnitSystem)[keyof typeof CalculationInputUnitSystem];

export const CalculationInputUnitSystem = {
  imperial: "imperial",
  metric: "metric",
} as const;

export interface CalculationInput {
  /** Optional profile ID to use as base */
  profileId?: number | null;
  /** FPS */
  muzzleVelocity: number;
  ballisticCoefficient: number;
  bcModel: CalculationInputBcModel;
  /** grains */
  bulletWeight: number;
  /** inches */
  bulletDiameter: number;
  /** yards */
  zeroRange: number;
  /** inches above bore */
  scopeHeight: number;
  /** Maximum range in yards to calculate */
  maxRange: number;
  /** Distance step between calculation points in yards */
  rangeStep: number;
  /** Wind speed in mph */
  windSpeed: number;
  /** Wind angle in degrees (0=headwind, 90=full value right crosswind) */
  windAngle: number;
  /** Temperature in Fahrenheit */
  temperature: number;
  /** Altitude in feet above sea level */
  altitude: number;
  /** Relative humidity 0-100 */
  humidity: number;
  /** Barometric pressure in inHg */
  pressure: number;
  /** Angle to target in degrees (uphill positive, downhill negative) */
  targetAngle: number;
  unitSystem: CalculationInputUnitSystem;
}

export interface TrajectoryPoint {
  /** Range in yards */
  range: number;
  /** Bullet drop in inches */
  drop: number;
  /** Wind drift in inches */
  windDrift: number;
  /** Velocity in fps */
  velocity: number;
  /** Energy in ft-lbs */
  energy: number;
  /** Time of flight in seconds */
  timeOfFlight: number;
  /** Drop correction in MOA */
  dropMOA: number;
  /** Drop correction in MRAD (milliradians) */
  dropMRAD: number;
  /** Wind correction in MOA */
  windMOA: number;
  /** Wind correction in MRAD */
  windMRAD: number;
}

export interface CalculationResult {
  trajectory: TrajectoryPoint[];
  maxRange: number;
  zeroRange: number;
  /** Maximum range within +/- 3 inches of zero */
  pointBlankRange: number;
  /** Range where bullet goes subsonic (fps < 1340) */
  supersonicLimit: number;
}