# lib/api-zod/src/generated/types/calculationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ballistic Calculator API
 * OpenAPI spec version: 0.1.0
 */
import type { CalculationInputBcModel } from "./calculationInputBcModel";
import type { CalculationInputUnitSystem } from "./calculationInputUnitSystem";

export interface CalculationInput {
  /** Optional profile ID to use as base */
  profileId?: number | null;
  /** FPS */
  muzzleVelocity: number;
  ballisticCoefficient: number;
  bcModel: CalculationInputBcModel;
  /** grains */
  bulletWeight: number;
  /** inches */
  bulletDiameter: number;
  /** yards */
  zeroRange: number;
  /** inches above bore */
  scopeHeight: number;
  /** Maximum range in yards to calculate */
  maxRange: number;
  /** Distance step between calculation points in yards */
  rangeStep: number;
  /** Wind speed in mph */
  windSpeed: number;
  /** Wind angle in degrees (0=headwind, 90=full value right crosswind) */
  windAngle: number;
  /** Temperature in Fahrenheit */
  temperature: number;
  /** Altitude in feet above sea level */
  altitude: number;
  /** Relative humidity 0-100 */
  humidity: number;
  /** Barometric pressure in inHg */
  pressure: number;
  /** Angle to target in degrees (uphill positive, downhill negative) */
  targetAngle: number;
  unitSystem: CalculationInputUnitSystem;
}