# lib/api-zod/src/generated/types/calculationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ballistic Calculator API
 * OpenAPI spec version: 0.1.0
 */
import type { TrajectoryPoint } from "./trajectoryPoint";

export interface CalculationResult {
  trajectory: TrajectoryPoint[];
  maxRange: number;
  zeroRange: number;
  /** Maximum range within +/- 3 inches of zero */
  pointBlankRange: number;
  /** Range where bullet goes subsonic (fps < 1340) */
  supersonicLimit: number;
}