# lib/api-zod/src/generated/types/createProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ballistic Calculator API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProfileInputBcModel } from "./createProfileInputBcModel";

export interface CreateProfileInput {
  name: string;
  caliber: string;
  bulletWeight: number;
  bulletDiameter: number;
  muzzleVelocity: number;
  ballisticCoefficient: number;
  bcModel: CreateProfileInputBcModel;
  zeroRange: number;
  scopeHeight: number;
  rifleWeight: number;
  notes?: string | null;
}