# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ballistic Calculator API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileBcModel } from "./profileBcModel";

export interface Profile {
  id: number;
  name: string;
  caliber: string;
  /** Bullet weight in grains */
  bulletWeight: number;
  /** Bullet diameter in inches */
  bulletDiameter: number;
  /** Muzzle velocity in feet per second */
  muzzleVelocity: number;
  /** Ballistic coefficient (G1 or G7) */
  ballisticCoefficient: number;
  /** Drag model used for BC */
  bcModel: ProfileBcModel;
  /** Zero range in yards */
  zeroRange: number;
  /** Scope height above bore in inches */
  scopeHeight: number;
  /** Rifle weight in pounds (for recoil calc) */
  rifleWeight: number;
  notes?: string | null;
  createdAt: string;
}