# lib/api-zod/src/generated/types/trajectoryPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ballistic Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface TrajectoryPoint {
  /** Range in yards */
  range: number;
  /** Bullet drop in inches */
  drop: number;
  /** Wind drift in inches */
  windDrift: number;
  /** Velocity in fps */
  velocity: number;
  /** Energy in ft-lbs */
  energy: number;
  /** Time of flight in seconds */
  timeOfFlight: number;
  /** Drop correction in MOA */
  dropMOA: number;
  /** Drop correction in MRAD (milliradians) */
  dropMRAD: number;
  /** Wind correction in MOA */
  windMOA: number;
  /** Wind correction in MRAD */
  windMRAD: number;
}